+++ src/labels.rs
@@ -1,3 +1,5 @@
+use std::fmt::Display;
+
 use crate::{config::Config, error::Error, todoist};
 use serde::Deserialize;
 
@@ -5,8 +7,15 @@ use serde::Deserialize;
 pub struct Label {
     pub name: String,
 }
-pub async fn get_labels(config: &Config) -> Result<Vec<Label>, Error> {
-    todoist::labels(config).await
+
+impl Display for Label {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let name = self.name.clone();
+        write!(f, "{name}")
+    }
+}
+pub async fn get_labels(config: &Config, spinner: bool) -> Result<Vec<Label>, Error> {
+    todoist::labels(config, spinner).await
 }
 
 pub fn json_to_labels(json: String) -> Result<Vec<Label>, Error> {