+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased (on main branch only)
 
+- Use multi-select in `task create` when no flags are used
+- Standardized language used in inputs
+
 ## 2024-12-19 v0.6.18
 
 - Added due to multi-select in `task edit`